refactor: update index module
